add resources specific util
