+++ src/app/resources/view-resources/resources-viewer.component.ts
@@ -9,7 +9,8 @@ import { StateService } from '../../shared/state.service';
 import { UserService } from '../../shared/user.service';
 import { CouchService } from '../../shared/couchdb.service';
 import { CSV_PREVIEW_MAX_BYTES, CSV_PREVIEW_MAX_ROWS, CsvService } from '../../shared/csv.service';
-import { couchAttachmentPath, formatResourceAttachmentSize, resourceAttachmentFilename } from '../../shared/utils';
+import { couchAttachmentPath } from '../../shared/utils';
+import { formatResourceAttachmentSize, resourceAttachmentFilename } from '../resources.utils';
 import { NgClass, NgTemplateOutlet } from '@angular/common';
 import { MatIconButton, MatAnchor } from '@angular/material/button';
 import { MatIcon } from '@angular/material/icon';
